Check valid user before leaving

+++ collaboration-svc/controllers/room_controller.py
@@ -1,6 +1,7 @@
 import asyncio
 from asyncio import Event
 from controllers.websocket_controller import WebSocketManager
+from fastapi import HTTPException
 from redis.asyncio import Redis
 from services.redis_event_queue import (
     get_match_confirmation_event,
@@ -10,7 +11,7 @@ from services.redis_event_queue import (
 )
 from services.redis_room_service import create_room, get_partner, is_user_alive, cleanup, add_room_cleanup, get_room_id, check_room_cleanup, delete_user_ttl
 from utils.logger import log
-from utils.utils import acquire_lock, release_lock, get_envvar, format_user_room_key, extract_information_from_event, format_heartbeat_key, format_cleanup_key
+from utils.utils import acquire_lock, release_lock, get_envvar, format_user_room_key, extract_information_from_event, format_heartbeat_key, format_cleanup_key, does_key_exist
 
 LOCK_KEY = "event_manager_lock"
 
@@ -83,6 +84,13 @@ async def remove_user(user_id: str, room_connection: Redis, websocket_manager: W
     Removes the user from the room.
     """
     heartbeat_key = format_heartbeat_key(user_id)
+
+    if (not does_key_exist(heartbeat_key, room_connection)):
+        raise HTTPException(
+                status_code=400,
+                detail="Cannot leave the match as user does not belong to this room"
+            )
+
     await delete_user_ttl(heartbeat_key, room_connection)
     await check_empty_room(user_id, room_connection, websocket_manager)
